refactor(data models): change data model collection names
Change data model collection names to state and sockets

BREAKING CHANGE: Change folder names

re #5
