+++ libs/models/src/index.ts
@@ -1,6 +1,6 @@
-export * from './lib/app/player';
-export * from './lib/app/race';
-export * from './lib/app/tournament';
-export * from './lib/app/leaderboard';
-export * from './lib/app/stateModel';
-export * from './lib/app/errorLog';
+export * from './lib/data/player';
+export * from './lib/data/race';
+export * from './lib/data/tournament';
+export * from './lib/data/leaderboard';
+export * from './lib/data/stateModel';
+export * from './lib/data/errorLog';
