+++ libs/models/src/lib/data/stateModel.ts
