+++ libs/models/src/lib/data/tournament.ts
@@ -1,70 +1,21 @@
-import { Leaderboard } from './leaderboard';
-import { Player } from './player';
-import { Race } from './race';
+import { AppPlayerId } from './player';
+import { AppRaceId } from './race';
 
-export enum TournamentState {
-  /** **Lobby**
-   *
-   * Every players are in lobby.
-   * *(Race is not started yet.)*
-   *
-   */
+export enum AppTournamentState {
   Lobby = 'lobby',
-
-  /** **Ready**
-   *
-   * Two or more players available in the lobby.
-   * *(Race is not started yet.)*
-   */
   Ready = 'ready',
-
-  /** **Countdown**
-   *
-   * A player pressed the `Play` button.
-   */
   Countdown = 'countdown',
-
-  /** **Race**
-   *
-   * The race is ongoing.
-   */
   Race = 'race',
-
-  /** **Leaderboard**
-   *
-   * The race is finished.
-   */
   Leaderboard = 'leaderboard',
-
-  /** **Empty**
-   *
-   * Lobby has no players. Waiting to destroy the lobby.
-   */
   Empty = 'empty',
 }
 
-export interface Tournament {
-  /** Unique tournament id (/lobby ID) */
-  id: string;
+/** Sample Id - `T:skt_2JVn` */
+export type AppTournamentId = `T:${string}`;
 
-  /** Tournament state */
-  state: TournamentState;
-
-  /** Races history and currect race
-   *
-   * can be empty if no races are started yet.
-   */
-  races: Race[];
-
-  /** Players in the tournament
-   *
-   * can be empty if lobby is empty.
-   */
-  players: Player[];
-
-  /** Leaderboards of the tournament
-   *
-   * can be empty if first race is not finished yet.
-   */
-  raceLeaderboards: Leaderboard[];
+export interface AppTournament {
+  state: AppTournamentState;
+  raceIds: Array<AppRaceId>;
+  playerIds: Array<AppPlayerId>;
 }
+export type AppTournaments = Record<AppTournamentId, AppTournament>;
